% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{convert_string}
\alias{convert_string}
\title{Simple funtion which can takes a string and returns it converted to another type}
\usage{
convert_string(str, type)
}
\arguments{
\item{str}{A string.}

\item{type}{A string.}
}
\value{
A string
}
\description{
Simple funtion which can takes a string and returns it converted to another type
}
