% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{convertNAs}
\alias{convertNAs}
\title{Convert NAs to empty character strings}
\usage{
convertNAs(x)
}
\arguments{
\item{x}{A string vector.}
}
\value{
A string vector.
}
\description{
Convert NAs to empty character strings
}
