% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{eSetToMASV}
\alias{eSetToMASV}
\title{Write MASV file from ExpressionSet}
\usage{
eSetToMASV(eset, filename)
}
\arguments{
\item{eset}{A ExpressionSet.}

\item{filename}{A string.}
}
\description{
Write MASV file from ExpressionSet
}
