% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{getMASVVersion}
\alias{getMASVVersion}
\title{Get version number of MASV file}
\usage{
getMASVVersion(topLeft)
}
\arguments{
\item{topLeft}{A string.}
}
\value{
A number
}
\description{
Get version number of MASV file
}
