% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{getRType}
\alias{getRType}
\title{Get R type of given data type}
\usage{
getRType(masvDataType)
}
\arguments{
\item{masvDataType}{A string.}
}
\value{
A string
}
\description{
Get R type of given data type
}
