% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{multi_data_set_to_masv}
\alias{multi_data_set_to_masv}
\title{Write MASV file from MultiDataSet}
\usage{
multi_data_set_to_masv(multi, filename)
}
\arguments{
\item{multi}{A MultiDataSet.}

\item{filename}{A string.}
}
\description{
Write MASV file from MultiDataSet
}
