% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseCell}
\alias{parseCell}
\title{Parse data from a single cell}
\usage{
parseCell(cell_data, rType, col_num)
}
\arguments{
\item{cell_data}{A string.}

\item{rType}{A string.}

\item{col_num}{A integer.}
}
\value{
A string
}
\description{
Parse data from a single cell
}
