% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseData}
\alias{parseData}
\title{Parse a row of data.}
\usage{
parseData(cols, dataType)
}
\arguments{
\item{cols}{A string vector.}

\item{dataType}{A string.}
}
\value{
A string vector
}
\description{
Parse a row of data.
}
