% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseExpressionSets}
\alias{parseExpressionSets}
\title{Parse MASV file to list of ExpressionSets}
\usage{
parseExpressionSets(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
A list
}
\description{
Parse MASV file to list of ExpressionSets
}
