% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMultiDataSet}
\alias{parseMultiDataSet}
\title{Parse MASV file to MultiDataSet}
\usage{
parseMultiDataSet(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
A MultiDataSet
}
\description{
Parse MASV file to MultiDataSet
}
