% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMultiDatSet}
\alias{parseMultiDatSet}
\title{Parse MASV file to MultiDataSet}
\usage{
parseMultiDatSet(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
A MultiDataSet
}
\description{
Parse MASV file to MultiDataSet
}
